@@ -35,22 +35,17 @@
 # Contributors:
 # Adam Zsarnoczay
 
-"""This module assigns households to housing units using the pyncoda package."""
+"""
+This script generates the 'fips_lookup.json' file.
+
+It fetches a national FIPS code file from a URL, reads it using the csv module,
+filters it for county and state-level codes, and saves the result
+as a clean JSON lookup.
+"""
 
 import csv
 import json
 from pathlib import Path
-
-import pandas as pd
-from pandas.errors import EmptyDataError, ParserError
-
-"""
-This script generates the 'fips_lookup.json' file.
-
-It fetches a national FIPS code file from a URL, reads it into pandas,
-filters it for county and state-level codes, and saves the result
-as a clean JSON lookup.
-"""
 
 
 def create_fips_lookup(csv_filepath: str, json_filepath: str) -> None:
@@ -106,7 +101,7 @@ def create_fips_lookup(csv_filepath: str, json_filepath: str) -> None:
     except FileNotFoundError:
         print(f"Error: The file '{csv_filepath}' was not found.")
         return
-    except (EmptyDataError, ParserError, OSError) as e:
+    except (csv.Error, OSError) as e:
         print(f'An error occurred: {e}')
         return
 
